refactored handlers
